move processNumberValue to its own file
